change SQL id types from INT to BIGINT (to match Java id long type)

+++ library/src/main/resources/schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE author
 (
-    author_id INT PRIMARY KEY AUTO_INCREMENT,
+    author_id BIGINT PRIMARY KEY AUTO_INCREMENT,
     firstname varchar NOT NULL,
     lastname  varchar NOT NULL,
     homeland  varchar,
@@ -11,7 +11,7 @@ CREATE TABLE author
 
 CREATE TABLE genre
 (
-    genre_id INT PRIMARY KEY AUTO_INCREMENT,
+    genre_id BIGINT PRIMARY KEY AUTO_INCREMENT,
     name     varchar NOT NULL,
     CONSTRAINT genre_unique UNIQUE (name)
 );
@@ -20,9 +20,9 @@ CREATE TABLE genre
 
 CREATE TABLE book
 (
-    book_id   INT PRIMARY KEY AUTO_INCREMENT,
-    author_id INT     NOT NULL REFERENCES author (author_id) ON DELETE CASCADE,
-    genre_id  INT REFERENCES genre (genre_id) ON DELETE CASCADE,
+    book_id   BIGINT PRIMARY KEY AUTO_INCREMENT,
+    author_id BIGINT  NOT NULL REFERENCES author (author_id) ON DELETE CASCADE,
+    genre_id  BIGINT REFERENCES genre (genre_id) ON DELETE CASCADE,
     title     varchar NOT NULL,
     year      int,
     CONSTRAINT book_unique UNIQUE (author_id, title, year)
